requirements.txt
# FastAPI framework
fastapi==0.103.1

# ASGI server for serving FastAPI
uvicorn==0.23.0

# Database interaction with SQLAlchemy
sqlalchemy==1.4.46
sqlalchemy-utils==0.38.3

# Pydantic for data validation
pydantic==2.7.0

# JWT library for token generation and validation
python-jose==3.3.0

# Password hashing
passlib[bcrypt]==1.7.4
# Email handling
python-dotenv==1.0.0
#smtplib

# For handling test cases
pytest==7.4.0

# Optional: If you plan to use Alembic for migrations
alembic==1.12.0

# Any async database support (e.g., async SQLite or PostgreSQL)
databases==0.7.0

pydantic-settings==2.7.0

python-jose==3.3.0

PyJWT==2.10.1